requirements.txt
Flask>=2.1.0
Werkzeug>=2.1.0
paho-mqtt==1.6.1
PyYAML==5.4.1
requests==2.26.0
Flask-Bcrypt>=1.0.0
Flask-Login>=0.6.0
pyotp>=2.6.0
qrcode[pil]>=7.3
pytz
cryptography
redis>=4.0.0